createMetProject

createMetProject Creates basic aermet project, specifying the directory, surface station (surf), onsite station (on), upper air (ua) station, landuse data, start date and ending date.

The function is designed to perform basic checks to make sure that the data are valid (e.g. WBAN and USAF id's match http://www1.ncdc.noaa.gov/pub/data/noaa/isd-history.txt)

The function also uses the make.names function to save the rmet class object as an .rds file in the project_Dir

This can be re-read into the R session with readRDS.

createMetProject(
  project_Name,
  project_Dir,
  start_Date,
  end_Date,
  surf_WBAN,
  surf_USAF,
  surf_Call,
  surf_Latitude,
  surf_Longitude,
  surf_Elevation,
  surf_AnenometerHeight,
  surf_UTC,
  ifg,
  ua_WMO,
  ua_WBAN = NULL,
  ua_UTC,
  ua_Latitude,
  ua_Longitude,
  lc_File,
  lc_Type,
  imp_File = NULL,
  imp_Type = NULL,
  cnpy_File = NULL,
  cnpy_Type = NULL,
  as_Snow,
  as_Arid,
  as_Moisture,
  as_Airport,
  as_Winter_NS,
  as_Winter_WS,
  as_Spring,
  as_Summer,
  as_Autumn,
  as_nsector = NULL,
  as_radius = 1,
  nws_sector = data.frame(index = 1:12, start = seq(0, 330, 30), end = seq(30, 360,
    30), ap = rep("AP", 12)),
  os_sector = data.frame(index = 1:12, start = seq(0, 330, 30), end = seq(30, 360, 30),
    ap = rep("NONAP", 12)),
  onsite_Latitude = NULL,
  onsite_Longitude = NULL,
  onsite_Fstring = NULL
)